Issue #75: Move point detectors to calibration folder.
