+++ sksurgeryimage/calibration/chessboard_point_detector.py
@@ -8,7 +8,7 @@ import logging
 import copy
 import cv2
 import numpy as np
-from sksurgeryimage.processing.point_detector import PointDetector
+from sksurgeryimage.calibration.point_detector import PointDetector
 
 LOGGER = logging.getLogger(__name__)
