+++ sksurgeryimage/calibration/point_detector.py
